+++ R/get_oews.R
@@ -114,3 +114,135 @@ get_oews <- function(simplify_table = TRUE, suppress_warnings = TRUE, return_dia
   
   return(result)
 }
+
+#' Download OEWS Area Definitions
+#'
+#' @param ref_year Four-digit year (converted to integer). The year for which to retrieve OEWS area definitions. Valid values are 2024 through current release year. Prior years included Township codes, which change the structure of the file.
+#' @param silent Logical. If TRUE (default), suppress console output
+#' @param geometry Logical.  If TRUE (default), downloads shapefiles for OEWS area definitions using `tigris::counties()` and `tigris::shift_geometry()` to render Alaska, Hawaii, and Puerto Rico with a focus on the area of the continental United States.
+#' 
+#' @return Data table which maps individual counties to OEWS area definitions.
+#'   \itemize{
+#'     \item fips_code - The State FIPS code
+#'     \item state_name - The state name
+#'     \item state_abb - The state two-character postal abbreviation
+#'     \item oews_area_code - The OEWS area code defining the metropolitan area or nonmetropolitan area the county belongs to.
+#'     \item oews_area_name - The OEWS area name
+#'     \item county_code - The FIPS code for the county
+#'     \item county_name - The county name
+#'     }
+#' 
+#' @export
+#' 
+#' @importFrom httr GET
+#' @importFrom httr write_disk
+#' @importFrom httr add_headers
+#' @importFrom readxl read_excel
+#' @importFrom tigris counties
+#' @importFrom sf st_union
+#' @importFrom dplyr select
+#' @importFrom dplyr left_join
+#' @importFrom dplyr group_by
+#' @importFrom dplyr summarize
+#' 
+#' @examples
+#' \donttest{
+#'  # Get OEWS area definitions without shapefiles and with processing messages.
+#'  test <- get_oews_areas(ref_year = 2024, geometry = FALSE, silent = FALSE)
+#'  
+#' }
+#' 
+get_oews_areas <- function(ref_year, silent = TRUE, geometry = TRUE){
+  
+  # Validate ref_year input
+  current_year <- as.integer(format(Sys.Date(), "%Y"))
+  min_year <- 2024
+  max_year <- current_year - 1
+  
+  if (is.na(as.integer(ref_year)) || length(ref_year) != 1) {
+    stop("`ref_year` must be coercable to a single integer value.")
+  }
+  
+  dl_year <- as.integer(ref_year)
+  
+  if (dl_year < min_year || dl_year > max_year) {
+    stop(sprintf("`ref_year` must be between %d and %d. Estimates are generally released in April for the prior year.", min_year, max_year))
+  }
+  
+  # Create download URL
+  oews_url <- paste0("https://www.bls.gov/oes/",dl_year,"/may/area_definitions_m",dl_year,".xlsx")
+  
+  headers <- c(
+    "Accept" = "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7",
+    "Accept-Encoding" = "gzip, deflate, br",
+    "Accept-Language" = "en-US,en;q=0.9",
+    "Connection" = "keep-alive",
+    "Host" = "www.bls.gov",
+    "Referer" = "https://www.bls.gov/oes/",
+    "Sec-Ch-Ua" = 'Not_A Brand";v="8", "Chromium";v="120", "Google Chrome";v="120"',
+    "Sec-Ch-Ua-Mobile" = "?0",
+    "Sec-Ch-Ua-Platform" = '"Windows"',
+    "Sec-Fetch-Dest" = "document",
+    "Sec-Fetch-Mode" = "navigate",
+    "Sec-Fetch-Site" = "same-origin",
+    "Sec-Fetch-User" = "?1",
+    "Upgrade-Insecure-Requests" = "1",
+    "User-Agent" = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36"
+  )
+  
+  # Download Excel file
+  if(!silent){
+    message("Downloading OEWS area definitions from BLS.")
+  }
+  response <- httr::GET(oews_url, 
+                        httr::write_disk(tf <- tempfile(fileext = ".xlsx")), 
+                        httr::add_headers(.headers = headers))
+  
+  # Check for successful response
+  httr::stop_for_status(response)
+  
+  # Track processing steps
+  processing_steps <- character(0)
+  
+  # Read and process Excel file
+  if(!silent){
+    message(paste0("Processing OEWS area definition Excel file for ",dl_year,"."))
+  }
+  oews_areas <- readxl::read_excel(
+    tf,
+    skip = 1,
+    col_types = c("text", "text", "text", "text", "text", "text", "text"),
+    col_names = c("fips_code", "state_name", "state_abb", "oews_area_code", "oews_area_name", "county_code", "county_name")
+  ) |> 
+    dplyr::mutate(
+      oews_area_code = stringr::str_pad(oews_area_code, width = 7, side = "left", pad = "0")
+    )
+  
+  # Clean up temporary file
+  unlink(tf)
+  
+  if(geometry == TRUE){
+    
+    if(!silent){
+      message("Creating OEWS shapefiles")
+    }
+    
+    oews_area <- oews_areas |> 
+      dplyr::mutate(GEOID = paste0(fips_code,county_code)) |>
+      dplyr::select(GEOID, oews_area_code, oews_area_name)
+    
+    area_shapes <- tigris::counties(year = dl_year, progress_bar = FALSE) |> 
+      tigris::shift_geometry() |>
+      dplyr::select(GEOID, geometry)
+    
+    oews_areas <- area_shapes |>
+      dplyr::left_join(oews_area, by = "GEOID") |>
+      dplyr::group_by(oews_area_name, oews_area_code) |>
+      dplyr::summarize(geometry = sf::st_union(geometry), .groups = "drop")
+  }
+  
+  
+  return(oews_areas)
+  
+  
+}