+++ NEWS.md
@@ -1,3 +1,8 @@
-# BLSloadR 0.1.5
+# BLSloadR 0.2
 
-* Initial CRAN submission.
+* Initial CRAN release.
+
+# BLSloadR 0.2.1
+
+* Additional web article on using OEWS data. (in progress)
+* Add helper function to download OEWS area definitions from BLS.